refactor: split widgets container into card / print / deck-button files
Reintroduces the split that landed in f6bc814d and was rolled back in
3e6caf42. Differences from the original split (defensive against the
Turbopack chunk-evaluation TDZ that surfaced in CI):

- Use function declarations with `export default function` instead of
  `const Component = ...; export default memo(Component)` — hoisted
  bindings are immune to cross-chunk TDZ regardless of how Turbopack
  splits modules.
- Drop the memo() wrapper around WidgetCard. The parent's widgetsAvailable
  array is already useMemo'd, so each render already gets a stable widget
  reference; adding memo here doesn't change render behavior and only
  introduced an extra module-evaluation reference.
